(viz) debug. colormap max and min values got mixed. take into account NaN values.
When figuring out colormap limits for negative and positive fields,
the maximum and minimum values were mistakenly identified as
minimum and maximum values, respectively.

Also, in checking for positive and negative definite fields, NaN values
were not taken into account, resulting in a positive definite field with
NaN values wrongly identified as a positive and negative field.  Stacking
the DataFrame of the field removes any NaN values, so doing this before
checking for positive/negative definiteness corrects this.