(viz) get contour levels such that zero is centred between two of the levels
Suppose you want to divide the range (vmin0, vmax0) into roughly Nsteps0 intervals,
where vmin0 < 0 < vmax0.  This function returns a new range(vmin, vmax) and new Nsteps that is
as close to the desired one as possible, but are such that dv = (vmax - vmin) / Nsteps
is a nice round value, and the value zero is centred on one of the Nsteps intervals.