Make nice datetime axis tick locations and labels
When the plotted values are datetime labels, i.e. of
type np.datetime64, this new feature will place ticks
along the axis in a way such that they do not crowd
the axis, and the tick labels are such that they do not
repeat for common time scales.  For example, if 2 adjacent
ticks are at 2012-10-11 and 2012-10-12, then only '11' and '12'
are placed at the ticks, but '2012-10' will be placed at only one
of them.