Be able to load either string or stringIO
When loading data, if a string is provided as,
say, Model.fpath_flux, then it will be assumed
to be the file path to the data UNDER the directory
Model.rundir.  However, if a StringIO is provided,
then it will assumed to be the content of the opened
file containing the data.