Nbconvert to script some procedures for lblnew
lblnew.f is the code for investigating a new
k-ditribution method for atmosphere radiation.

Because we are trying to fit this parameterisation
by hand, many cases (sets of parameter values) need
to be submitted by hand.  A lot of this process is
repetitive.

Here we have written some procedures that link up
different stages in this process so that it is
more automated.

The main steps for each case are:

(1) Put all input values in a Python dictionary
(2) Create a run directory, using the input values, for
    the case.
(2) Copy lblnew.f to the run directory.
(3) Insert the input values into lblnew.f.
(4) Compile lblnew.f and execute.
(5) After lblnew.f has finished running, create
    an analysis directory for the case.
(6) Copy result.ipynb to the analysis directory.
(7) Create and fill params.py in the analysis
    directory.
(8) Execute the analysis notebook.
(9) After the analysis is finished.  Add and commit
    the 'params.py' and 'results.ipynb'.

Steps (1) to (8) for multiple cases can be run in parallel,
using Python's subprocess module.