Copy over 'results.ipynb' template notebook
This is a template notebook for presenting results
from lblnew.f.  It makes more sense that this template
is developed alongside the python functions used in it,
here in this repository, rather than the repository that
contains the notebooks that have been run for different
cases.  This template notebook is to be used for all cases,
but its purpose is not to be shown to other people.