Create pipeline module
This module will mainly use the subprocess module
to parallelise the execution of the fortran code
(lblnew.f), followed by the execution of the analysis
notebook (.ipynb), then the committing of the resulting
notebook to Git repository.

It can be provided with different versions of the fortran
code, or the notebook.