Create new module for kdist bestfit runs
The 'kdist bestfit runs' are the runs of the version
of 'lblnew.f' in which values for k-distribution method's
parameters are set as input, such as `ng_refs`.  This version
of 'lblnew.f' is only applicable for a single molecule, and
really only makes sense for a single spectral band.  The
purpose of running this version is to empirically find the
optimal values for the k-distribution method's parameters.