Create module for running overlap cases
'overlap cases' uses versions of 'lblnew.f' that
calculate radiation when one or more absorber
may be present.  In these calculations, the
k-dist method parameters are assumed to be hard-wired
in lblnew.f; these would most likely be values that
have been empirically determined to be optimal for
absorbers in particular spectral bands.

So, there are fewer input parameters to be specified, because
parameters like `ng_refs`, `wgt` are already hard-wired in
'lblnew.f'.