Run analysis notebook given input parameter values
Created function runs the analysis after clirad-lw has
been run.  It creates a directory for the analysis, copies
the template notebook to the directory, creates a 'param.py'
containing the path to the corresponding fortran run and
the input parameter values, copies that to the directory,
then it executes the notebook.