Update to crd-d2f6fe2 and compute k-table
This is a major round of k-table computation.
The k-dist parameter values have been manually
adjusted several times and are believed to be
best-fit values.  The lblnew-bestfit code in
the crd repository can also be considered stable.

Comment
-------
At this moment, the subroutine transf() in crd is
still being fixed because it produces unwanted spikey
features in its output, for `igg=1`.  However,
this does not affect the k-tables computed.