Change regex for parameter input
The previous way of enterting input into the
Fortran code failed when h2o concentration is
implicitly 0.  Simple string.replace also replaces
the variable name with a number, resulting in
expressions like '0. = 0.', when it should be
'wa(i, k) = 0.'.

In this commit, different regex groups are used to
search for where input are to be entered, and instead
of using string.replace(), re.sub, with back reference
to matched groups, is used to enter the desired value.