Update n2o band6 parameters
Prof Chou:
The idea is to use only one "g" to compute the absorption
of n2o in Band 6 (1100-1215). For a very weak absorption
band, the  absorption coefficient averaged over a band
is only dependent upon temperature but nearly
independent of pressure. With this in mind, we want to
compute fluxes in the n2o Band 6 using a constant
absorption coefficient with a slight adjustment by temperature.