Pass dummy array if there is not data
If there is no data in the .dat file read
(i.e. if it's empty for some reason), use
a dummy array of 0s, and print out the path
 of the data file that is empty, so we know
the lblnew-bestfit case that needs to be re-run.
This is the main purpose of this change.

However, the dummy array of 0s is still nevertheless
written into Fortran, and I don't know what would
happen if this fortran is imported into clirad.
This should not be done anyway.