Create classes for clirad-lw, lblnew-bestfit/overlap parameters
Using dictionaries to hold the parameter values is good, but it
doesn't restrict what parameters are allowed for a model, and
since each model has a different set of parameter, it could be
a bit confusing and unsafe.  Now, if you try to create a set of
parameters for a model, but one of the parameters you specify
isn't actually one of the model's parameters, an Exception will
be raised.  For example, if you try to create a lblnew-overlap
parameter object with a `ng_refs` value, it will stop you because
`ng_refs` is a parameter for lblnew-bestfit, not lblnew-overlap.