Return parameter for overlapped cases
So, [(band 1, h2o), (band 3, co2)] is equal to
(band 1, h2o) + (band 3, co2), because the spectral bands
are different.  However, [(band 1, h2o), (band 1, co2)] is NOT
equal to (band 1, h2o) + (band 1, co2), because h2o and co2 are
in the same spectral band, so their overlapping needs to be taken
care of.  In this case, we need to used the overlapped run
(band 1, h2o + co2).