Always return to the directory in which the notebook is running
When building the fortran code, cd into the specified path first,
build, then return to the initial directory.  Similarly for submitting
the code using bsub: cd into the specified path first, then return
back to the initial directory.  The initial directory is usually the
directory in which the notebook is running from.