Test inputting into analysis notebook after it's loaded by nbformat
Before, to assign variables in the analysis notebook, you need
to first save the variables in file and place the file in the
same directory as the analysis notebook.  Then, inside the analysis
notebook, you then need to load the variables from that file.

In this commit, the analysis notebook is loaded into our notebook from
which we execute the analysis.  The analysis notebook is loaded into json
format.  We can then change the appropriate cell inputs in this json to load
the variables the analysis needs, such as the input parameter dictionary,
or the path to the fortran run, where the output data files are.